src/app/feature/scenarios/services/new-scenario-steps.guard.ts
import { Injectable } from '@angular/core';
import { ActivatedRouteSnapshot, CanActivateChild, RouterStateSnapshot, UrlTree } from '@angular/router';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root',
})
export class NewScenarioStepsGuard implements CanActivateChild {
  canActivateChild(
    childRoute: ActivatedRouteSnapshot,
    state: RouterStateSnapshot
  ): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
    const id = childRoute.paramMap.get('id') ?? childRoute.parent?.paramMap.get('id');
    const result = id !== 'new' || childRoute.routeConfig?.path === '';
    return result;
  }
}
